requirements.txt
fastapi
uvicorn
anthropic
pymupdf
pandas
openpyxl
streamlit
python-multipart
python-dotenv